Fix SlopeDiv overflow
Fixes issue #383. I've noticed similar anomalies in other maps with very large open areas (nuts.wad for instance).

+++ src/r_main.c
@@ -186,12 +186,12 @@ int R_PointOnSegSide(fixed_t x, fixed_t y, seg_t *line)
 
 int SlopeDiv(unsigned int num, unsigned int den)
 {
-    unsigned int    ans;
+    uint64_t    ans;
 
     if (den < 512)
         return SLOPERANGE;
 
-    ans = (num << 3) / (den >> 8);
+    ans = (int)((uint64_t)num << 3) / (den >> 8);
     return (ans <= SLOPERANGE ? ans : SLOPERANGE);
 }
